signout added to navigation bottom bar

+++ app/src/main/java/com/example/autogallerykotlin/ui/view/MainActivity.kt
@@ -1,15 +1,16 @@
+
 package com.example.autogallerykotlin.ui.view
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
 import androidx.navigation.NavController
+import androidx.navigation.findNavController
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.NavigationUI.setupWithNavController
 import com.example.autogallerykotlin.R
 import com.example.autogallerykotlin.databinding.ActivityMainBinding
+import com.example.autogallerykotlin.ui.fragment.AddAdvertiseFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -24,39 +25,28 @@ class MainActivity : AppCompatActivity() {
         setContentView(binding.root)
 
         binding.bottomNavigationView.background = null
-        binding.bottomNavigationView.menu.getItem(2).isEnabled = false
 
         val navHostFragment = supportFragmentManager.findFragmentById(R.id.fragmentContainerView) as NavHostFragment
         navController = navHostFragment.navController
 
         val bottomNavigationView = binding.bottomNavigationView
-        setupWithNavController(bottomNavigationView,navController)
+        setupWithNavController(bottomNavigationView, navController)
 
-    }
-/*
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        binding.bottomNavigationView.setOnItemSelectedListener {
 
-        val inflater = menuInflater
-        inflater.inflate(R.menu.bottom_nav_menu,menu)
-        return true
-    }
+            when (it.itemId) {
 
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+                R.id.SignOut -> {
+                    val sharedPreferences = getSharedPreferences("login", MODE_PRIVATE)
+                    val editor = sharedPreferences.edit()
+                    sharedPreferences.getString("users_id", null)
+                    editor.clear()
+                    editor.apply()
 
-        return when(item.itemId){
-            R.id.SignOut->{
-                val sharedPreferences = getSharedPreferences("login", MODE_PRIVATE)
-                val editor = sharedPreferences.edit()
-                sharedPreferences.getString("users_id",null)
-                editor.clear()
-                editor.apply()
-
-                startActivity(Intent(this, LoginActivity::class.java))
-                true
+                    startActivity(Intent(this, LoginActivity::class.java))
+                }
             }
-            else->{
-                super.onOptionsItemSelected(item)
-            }
+            true
         }
-    }*/
+    }
 }
